feat: add Modal component for project deletion confirmation and integrate with ProjectContent
