+++ src/components/index.js
@@ -7,3 +7,4 @@ export { default as ProjectList } from "./ProjectList";
 export { default as ProjectFilter } from "./ProjectFilter";
 export { default as ProjectContent } from "./ProjectContent";
 export { default as ProjectChat } from "./ProjectChat";
+export { default as Modal } from "./Modal";
